feat(admin): 添加二手书回收功能及界面优化 - 新增二手书回收菜单项和徽章显示功能 - 实现管理后台侧边栏头部标题区域 - 添加状态徽章实时刷新机制 - 优化样式主题和字体配置 - 集成支付宝回调处理二手订单支付逻辑 - 添加图书详情页异常处理和用户认证检查 - 优化图书搜索查询条件匹配逻辑 - 修复购物车添加操作的错误处理 - 更新应用根组件样式变量配置

+++ store-pojo/src/main/java/com/PTU/vo/BookVO.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 
 @Data
 public class BookVO {
-    private Integer id; // 主键ID
+    private Long id; // 主键ID（与 Book.id 一致，避免 Long→Integer 拷贝异常）
 
     private String title; // 书名
     private String author; // 作者
